Refactoring of content entities
1)Removed VerifiedContent. Instead of having a separate entity to represent a verified content we will be using the
UnverifiedContent(now named Contet) to represent it.
2)Renamed UnverifiedStudio,UnverifiedAnime,UnverifiedCharacter,UnverifiedNovel to StudioContent,AnimeContent,CharacterContent,NovelContent.
Also, added to each of them an id property because now ContentService no longer returns content entity so that the calling site can get the id,
instead the returned content's id is set(which technically is the same id that is passed as argument).
3)Refactored init.sql according to changes above.
4)Fixed insert_anime.html that used a message key that did not exist.