Refactor ViewController and layout template
1)Refactored ApplicationInitializer to properly insert an anime admin in to
the database for testing.
2)Refactored ViewController methods so that each calls a method that
initializes model attribute with name roles to contain the user roles or
empty list(in case user is not authenticated).
This is done because thymeleaf extra spring security 6 library doesn't
support hasRole and other expressions for spring webflux. So this is a workaround.
3)Fixed JwtServiceImpl which didn't join roles properly and was causing users to have an additional empty
role in scope claim.
4)Refactored layout.html to use roles attribute.