Fix @ModelAttribute method not being called before @RequestMapping
1)Refactored ViewController by changing roleAttributeInitializer method
implementation to use reactor type. When using suspend kotlin function, the
method is not invoked properly(it's invoked after @RequestMapping, and not
before as it should). So instead of calling the method in each @RequestMapping
method, I am using reactor types which do not cause any issues and the method
is invoked properly.
2)Also, fixed README which had ktor mentioned which isn't used anymore.