Add integration tests for ContentService methods related to studio content
1)Added integration tests for ContentService for methods related to studio content.
2)Refactored ContentService common verification code. Instead of throwing ISE when content is not under verification,
the verifier id is used to check if it equals to the caller which also covers verifier of the content being null.
This also tides the security of the method since the caller cannot know why the AccessDeniedException was thrown -
because the caller is not verifier or because there is not verifier.
2)Refactored AnimeContentServiceITest because of (2) change.